fix fatal error in frontend
